refactor(variant index): sorting out new variant index dataset
