--- src/gentropy/assets/schemas/variant_annotation.json
@@ -1,215 +1,0 @@
-{
-  "type": "struct",
-  "fields": [
-    {
-      "name": "variantId",
-      "type": "string",
-      "nullable": false,
-      "metadata": {}
-    },
-    {
-      "name": "chromosome",
-      "type": "string",
-      "nullable": false,
-      "metadata": {}
-    },
-    {
-      "name": "position",
-      "type": "integer",
-      "nullable": false,
-      "metadata": {}
-    },
-    {
-      "name": "referenceAllele",
-      "type": "string",
-      "nullable": false,
-      "metadata": {}
-    },
-    {
-      "name": "alternateAllele",
-      "type": "string",
-      "nullable": false,
-      "metadata": {}
-    },
-    {
-      "name": "chromosomeB37",
-      "type": "string",
-      "nullable": true,
-      "metadata": {}
-    },
-    {
-      "name": "positionB37",
-      "type": "integer",
-      "nullable": true,
-      "metadata": {}
-    },
-    {
-      "name": "alleleType",
-      "type": "string",
-      "nullable": true,
-      "metadata": {}
-    },
-    {
-      "name": "rsIds",
-      "type": {
-        "type": "array",
-        "elementType": "string",
-        "containsNull": true
-      },
-      "nullable": true,
-      "metadata": {}
-    },
-    {
-      "name": "alleleFrequencies",
-      "type": {
-        "type": "array",
-        "elementType": {
-          "type": "struct",
-          "fields": [
-            {
-              "name": "populationName",
-              "type": "string",
-              "nullable": true,
-              "metadata": {}
-            },
-            {
-              "name": "alleleFrequency",
-              "type": "double",
-              "nullable": true,
-              "metadata": {}
-            }
-          ]
-        },
-        "containsNull": true
-      },
-      "nullable": false,
-      "metadata": {}
-    },
-    {
-      "name": "inSilicoPredictors",
-      "nullable": false,
-      "metadata": {},
-      "type": {
-        "type": "struct",
-        "fields": [
-          {
-            "name": "cadd",
-            "nullable": true,
-            "metadata": {},
-            "type": {
-              "type": "struct",
-              "fields": [
-                {
-                  "name": "raw",
-                  "type": "float",
-                  "nullable": true,
-                  "metadata": {}
-                },
-                {
-                  "name": "phred",
-                  "type": "float",
-                  "nullable": true,
-                  "metadata": {}
-                }
-              ]
-            }
-          },
-          {
-            "name": "revelMax",
-            "type": "double",
-            "nullable": true,
-            "metadata": {}
-          },
-          {
-            "name": "spliceaiDsMax",
-            "type": "float",
-            "nullable": true,
-            "metadata": {}
-          },
-          {
-            "name": "pangolinLargestDs",
-            "type": "double",
-            "nullable": true,
-            "metadata": {}
-          },
-          {
-            "name": "phylop",
-            "type": "double",
-            "nullable": true,
-            "metadata": {}
-          },
-          {
-            "name": "siftMax",
-            "type": "double",
-            "nullable": true,
-            "metadata": {}
-          },
-          {
-            "name": "polyphenMax",
-            "type": "double",
-            "nullable": true,
-            "metadata": {}
-          }
-        ]
-      }
-    },
-    {
-      "name": "vep",
-      "type": {
-        "type": "struct",
-        "fields": [
-          {
-            "name": "mostSevereConsequence",
-            "type": "string",
-            "nullable": true,
-            "metadata": {}
-          },
-          {
-            "name": "transcriptConsequences",
-            "type": {
-              "type": "array",
-              "elementType": {
-                "type": "struct",
-                "fields": [
-                  {
-                    "name": "aminoAcids",
-                    "type": "string",
-                    "nullable": true,
-                    "metadata": {}
-                  },
-                  {
-                    "name": "consequenceTerms",
-                    "type": {
-                      "type": "array",
-                      "elementType": "string",
-                      "containsNull": true
-                    },
-                    "nullable": true,
-                    "metadata": {}
-                  },
-                  {
-                    "name": "geneId",
-                    "type": "string",
-                    "nullable": true,
-                    "metadata": {}
-                  },
-                  {
-                    "name": "lof",
-                    "type": "string",
-                    "nullable": true,
-                    "metadata": {}
-                  }
-                ]
-              },
-              "containsNull": true
-            },
-            "nullable": true,
-            "metadata": {}
-          }
-        ]
-      },
-      "nullable": false,
-      "metadata": {}
-    }
-  ]
-}
