--- src/gentropy/assets/schemas/variant_annotation_new.json
@@ -1,223 +1,0 @@
-{
-  "fields": [
-    {
-      "metadata": {},
-      "name": "variantId",
-      "nullable": true,
-      "type": "string"
-    },
-    {
-      "metadata": {},
-      "name": "chromosome",
-      "nullable": true,
-      "type": "string"
-    },
-    {
-      "metadata": {},
-      "name": "position",
-      "nullable": true,
-      "type": "integer"
-    },
-    {
-      "metadata": {},
-      "name": "referenceAllele",
-      "nullable": true,
-      "type": "string"
-    },
-    {
-      "metadata": {},
-      "name": "alternateAllele",
-      "nullable": true,
-      "type": "string"
-    },
-    {
-      "metadata": {},
-      "name": "inSilicoPredictors",
-      "nullable": false,
-      "type": {
-        "containsNull": true,
-        "elementType": {
-          "fields": [
-            {
-              "metadata": {},
-              "name": "method",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "assessment",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "score",
-              "nullable": true,
-              "type": "float"
-            },
-            {
-              "metadata": {},
-              "name": "assessmentFlag",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "targetId",
-              "nullable": true,
-              "type": "string"
-            }
-          ],
-          "type": "struct"
-        },
-        "type": "array"
-      }
-    },
-    {
-      "metadata": {},
-      "name": "mostSevereConsequenceId",
-      "nullable": true,
-      "type": "string"
-    },
-    {
-      "metadata": {},
-      "name": "transcriptConsequences",
-      "nullable": true,
-      "type": {
-        "containsNull": false,
-        "elementType": {
-          "fields": [
-            {
-              "metadata": {},
-              "name": "variantConsequenceIds",
-              "nullable": true,
-              "type": {
-                "containsNull": true,
-                "elementType": "string",
-                "type": "array"
-              }
-            },
-            {
-              "metadata": {},
-              "name": "amino_acid_change",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "uniprotAccessions",
-              "nullable": true,
-              "type": {
-                "containsNull": true,
-                "elementType": "string",
-                "type": "array"
-              }
-            },
-            {
-              "metadata": {},
-              "name": "isEnsemblCanonical",
-              "nullable": false,
-              "type": "boolean"
-            },
-            {
-              "metadata": {},
-              "name": "codons",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "distance",
-              "nullable": true,
-              "type": "long"
-            },
-            {
-              "metadata": {},
-              "name": "targetId",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "impact",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "transcriptId",
-              "nullable": true,
-              "type": "string"
-            }
-          ],
-          "type": "struct"
-        },
-        "type": "array"
-      }
-    },
-    {
-      "metadata": {},
-      "name": "rsIds",
-      "nullable": true,
-      "type": {
-        "containsNull": true,
-        "elementType": "string",
-        "type": "array"
-      }
-    },
-    {
-      "name": "alleleFrequencies",
-      "type": {
-        "type": "array",
-        "elementType": {
-          "type": "struct",
-          "fields": [
-            {
-              "name": "populationName",
-              "type": "string",
-              "nullable": true,
-              "metadata": {}
-            },
-            {
-              "name": "alleleFrequency",
-              "type": "double",
-              "nullable": true,
-              "metadata": {}
-            }
-          ]
-        },
-        "containsNull": true
-      },
-      "nullable": false,
-      "metadata": {}
-    },
-    {
-      "metadata": {},
-      "name": "dbXrefs",
-      "nullable": true,
-      "type": {
-        "containsNull": true,
-        "elementType": {
-          "fields": [
-            {
-              "metadata": {},
-              "name": "id",
-              "nullable": true,
-              "type": "string"
-            },
-            {
-              "metadata": {},
-              "name": "source",
-              "nullable": true,
-              "type": "string"
-            }
-          ],
-          "type": "struct"
-        },
-        "type": "array"
-      }
-    }
-  ],
-  "type": "struct"
-}
